requirements.txt
# 🦅 Hive Fleet Obsidian: Gen 63 Stack (O.B.S.I.D.I.A.N.)

# --- O: Orchestrator (Navigator) ---
temporalio>=1.4.0       # Durable Execution (The Manager)

# --- B: Bus (Bridger) ---
nats-py>=2.6.0          # NATS JetStream Client

# --- S: Schema (Immunizer) ---
pydantic>=2.5.0         # The "Intent" Layer (SSOT)
pydantic-settings>=2.0.0 # Settings Management
instructor>=1.0.0       # Structured Output (Pydantic <-> LLM)

# --- I: Intelligence (Shaper) ---
langgraph>=0.0.10       # SOTA State Machines (PREY Loop)
langchain>=0.1.0        # Main LangChain Package
langchain-openai>=0.0.5 # OpenRouter/OpenAI Integration
dspy-ai>=2.5.0          # DSPy for Prompt Optimization

# --- D: Database (Assimilator) ---
lancedb>=0.5.0          # Vector Database (Cold Stigmergy)
asyncpg>=0.29.0         # Async Postgres Driver
psycopg2-binary>=2.9.9  # Sync Postgres Driver (for tools)
pgvector>=0.2.4         # PGVector Python Client
sqlalchemy>=2.0.0       # ORM (Optional but recommended)

# --- I: Interface (Sensors/Effectors) ---
mcp>=0.1.0              # Model Context Protocol (Core)

# --- A: Analytics (Observer) ---
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
langsmith>=0.1.0        # Tracing & Debugging (Essential)

# --- N: Nodes (Injector) ---
ray>=2.9.0              # Distributed Compute (The Swarm)

# --- The Forge (Evolution & Utils) ---
ribs>=0.5.0             # MAP-Elites (Quality-Diversity)
numpy>=1.26.0           # Math for evolution
python-dotenv>=1.0.0    # Config
pytest>=8.0.0           # Testing
pytest-bdd>=7.0.0       # Behavior Driven Development (Gherkin)
pytest-asyncio>=0.23.0  # Async testing
tiktoken>=0.5.0         # Token counting
PyYAML>=6.0.0           # Configuration
pre-commit>=3.5.0
openai>=1.0.0           # Direct OpenAI Client
duckduckgo-search>=5.0.0  # Real Web Search
numexpr>=2.8.0          # Safe Calculation
rich>=13.0.0            # Terminal UI
sentence-transformers>=2.2.2 # Local Embeddings (Free & Fast)
aiofiles>=23.2.1
